--- a/go/vt/sqlparser/ast.go
+++ b/go/vt/sqlparser/ast.go
@@ -2896,7 +2896,8 @@ type (
 	}
 
 	CountStar struct {
-		_ bool
+		Name string
+		_    bool
 		// TL;DR; This makes sure that reference equality checks works as expected
 		//
 		// You're correct that this might seem a bit strange at first glance.