--- a/go/vt/sqlparser/parse_test.go
+++ b/go/vt/sqlparser/parse_test.go
@@ -5901,8 +5901,8 @@ func TestOne(t *testing.T) {
 	testOne := struct {
 		input, output string
 	}{
-		input:  "",
-		output: "",
+		input:  "Select Count(*) from t",
+		output: "select Count(*) from t",
 	}
 	if testOne.input == "" {
 		return