--- a/go/vt/sqlparser/sql.go
+++ b/go/vt/sqlparser/sql.go
@@ -18798,7 +18798,7 @@ yydefault:
 		var yyLOCAL Expr
 //line sql.y:5989
 		{
-			yyLOCAL = &CountStar{OverClause: yyDollar[5].overClauseUnion()}
+			yyLOCAL = &CountStar{Name: yyDollar[1].str, OverClause: yyDollar[5].overClauseUnion()}
 		}
 		yyVAL.union = yyLOCAL
 	case 1163: