<!--
  Thank you for your contribution to the Vitess project.
  How to contribute: https://vitess.io/docs/contributing/
  Please first make sure there is an open Issue to discuss the feature/fix suggested in this PR.
  If this is a new feature, please mark the Issue as "RFC".
 -->

<!-- if this PR is Work in Progress please create it as a Draft Pull Request -->

## Description

<!-- A few sentences describing the overall goals of the pull request's commits. -->
<!-- If this is a bug fix and you think the fix should be backported, please write so. -->

Function names in select expressions are returning incorrect column names in Vitess. For example, if a user runs `SELECT Count(*) from t`, Vitess's column name in the output is `count(*)`, even though MySQL would output `Count(*)`. Notice the  caseing of the count keyword.

This PR adds the Name field to the CountStar ast struct to keep track of the original casing. I have verified from the plan outputs that Vitess is now using the correct casing for unsharded and simple sharded queries. 

I have added an e2e benchmark to run the `Count(*)` query multiple times.

Here are the benchmark results for the e2e test and unit tests against main - 

```
pkg: vitess.io/vitess/go/test/endtoend/vtgate/gen4
             │ bench_main_e2e │           bench_fix_e2e            │
             │     sec/op     │   sec/op     vs base               │
CountStar-10      22.41m ± 1%   22.78m ± 6%  +1.66% (p=0.000 n=20)
```

```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/vt/sqlparser
                                  │ bench_main_unit │           bench_fix_unit           │
                                  │     sec/op      │   sec/op     vs base               │
ParseTraces/django_queries.txt-10       761.0µ ± 1%   761.0µ ± 2%       ~ (p=0.738 n=20)
ParseTraces/lobsters.sql.gz-10          32.05m ± 1%   32.25m ± 1%  +0.64% (p=0.012 n=20)
ParseStress/sql0-10                     3.572µ ± 0%   3.600µ ± 1%  +0.77% (p=0.008 n=20)
ParseStress/sql1-10                     11.28µ ± 0%   11.34µ ± 0%  +0.58% (p=0.029 n=20)
Parse3/normal-10                        1.749m ± 0%   1.746m ± 0%       ~ (p=0.383 n=20)
Parse3/escaped-10                       2.781m ± 0%   2.766m ± 0%  -0.54% (p=0.004 n=20)
geomean                                 410.4µ        411.3µ       +0.21%

                                  │ bench_main_unit │            bench_fix_unit             │
                                  │      B/op       │     B/op      vs base                 │
ParseTraces/django_queries.txt-10      557.9Ki ± 0%   557.9Ki ± 0%       ~ (p=0.457 n=20)
ParseTraces/lobsters.sql.gz-10         15.65Mi ± 0%   15.65Mi ± 0%       ~ (p=0.172 n=20)
ParseStress/sql0-10                    1.392Ki ± 0%   1.392Ki ± 0%       ~ (p=1.000 n=20) ¹
ParseStress/sql1-10                    4.186Ki ± 0%   4.186Ki ± 0%       ~ (p=1.000 n=20) ¹
Parse3/normal-10                       2.716Ki ± 0%   2.716Ki ± 0%       ~ (p=0.231 n=20)
Parse3/escaped-10                      4.914Mi ± 0%   4.914Mi ± 0%       ~ (p=0.551 n=20)
geomean                                94.49Ki        94.49Ki       +0.00%

                                  │ bench_main_unit │            bench_fix_unit            │
                                  │    allocs/op    │  allocs/op   vs base                 │
ParseTraces/django_queries.txt-10       5.357k ± 0%   5.357k ± 0%       ~ (p=1.000 n=20) ¹
ParseTraces/lobsters.sql.gz-10          238.3k ± 0%   238.3k ± 0%   0.00% (p=0.037 n=20)
ParseStress/sql0-10                      30.00 ± 0%    30.00 ± 0%       ~ (p=1.000 n=20) ¹
ParseStress/sql1-10                      82.00 ± 0%    82.00 ± 0%       ~ (p=1.000 n=20) ¹
Parse3/normal-10                         58.00 ± 0%    58.00 ± 0%       ~ (p=1.000 n=20) ¹
Parse3/escaped-10                        300.0 ± 0%    300.0 ± 0%       ~ (p=1.000 n=20) ¹
geomean                                  616.0         616.0       +0.00%
```
## Related Issue(s)

<!-- List related issues and pull requests. If this PR fixes an issue, please add it using Fixes #????  -->

## Checklist

-   [ ] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [ ] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [ ] Tests were added or are not required
-   [ ] Did the new or modified tests pass consistently locally and on CI?
-   [ ] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->