@GuptaManan100 2024-06-13
Like I suspected, just adding the `Name` field isn't enough. The planner has to change too to respect this field as can be seen by the current output for a more complex sharded query - 
```
{
  "QueryType": "SELECT",
  "Original": "select Count(*) from user join user_extra where user.intcol + user_extra.incol < 5 limit 5",
  "Instructions": {
    "OperatorType": "Limit",
    "Count": "5",
    "Inputs": [
      {
        "OperatorType": "Aggregate",
        "Variant": "Scalar",
        "Aggregates": "sum_count_star(0) AS Count(*)",
        "Inputs": [
          {
            "OperatorType": "Projection",
            "Expressions": [
              "(*) * (*) as Count(*)"
            ],
            "Inputs": [
              {
                "OperatorType": "Join",
                "Variant": "Join",
                "JoinColumnIndexes": "L:0,R:0",
                "JoinVars": {
                  "user_intcol": 1
                },
                "TableName": "`user`_user_extra",
                "Inputs": [
                  {
                    "OperatorType": "Route",
                    "Variant": "Scatter",
                    "Keyspace": {
                      "Name": "user",
                      "Sharded": true
                    },
                    "FieldQuery": "select (*), `user`.intcol from `user` where 1 != 1 group by `user`.intcol",
                    "Query": "select (*), `user`.intcol from `user` group by `user`.intcol",
                    "Table": "`user`"
                  },
                  {
                    "OperatorType": "Route",
                    "Variant": "Scatter",
                    "Keyspace": {
                      "Name": "user",
                      "Sharded": true
                    },
                    "FieldQuery": "select (*) from user_extra where 1 != 1 group by .0",
                    "Query": "select (*) from user_extra where :user_intcol + user_extra.incol < 5 group by .0",
                    "Table": "user_extra"
                  }
                ]
              }
            ]
          }
        ]
      }
    ]
  },
  "TablesUsed": [
    "user.user",
    "user.user_extra"
  ]
}
```