@GuptaManan100 2024-06-13
I pushed a fix and everything seems to be working fine now. We just needed to take care of all the places that were instantiating a new CountStar struct