@arthurschreiber 2024-06-14
Is this only specific to `Count(*)` or also to other aggregate functions like `Avg(...)` where the vtgate currently performs a rewrite?